Put second motor things into elevator and took out of generic superstructure
